Move weather widget files into subfolder
